FeelTunes/requirements.txt
Flask
Flask-SQLAlchemy==3.0.0
flask_wtf
email-validator
Werkzeug==3.0.0
Flask-Login
Flask-Mail
itsdangerous
python-dotenv
flask_session


